importing spandsp 0.0.4pre8 (same as snapshot 20070908)

+++ src/noise.c
@@ -23,7 +23,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- * $Id: noise.c,v 1.17 2006/11/28 16:59:56 steveu Exp $
+ * $Id: noise.c,v 1.18 2007/09/02 14:25:51 steveu Exp $
  */
 
 /*! \file */
@@ -100,7 +100,7 @@ noise_state_t *noise_init_dbov(noise_state_t *s, int seed, float level, int clas
         /* Allow for the gain of the filter */
         rms *= 1.043f;
     }
-    s->rms = (int32_t) (rms*sqrt(12.0/quality));
+    s->rms = (int32_t) (rms*sqrtf(12.0f/quality));
     s->class_of_noise = class_of_noise;
     return s;
 }
